requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
factory-boy==3.3.1
Faker==33.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tomli==2.1.0
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
